Fix ROUND() migration + remove undici dispatcher from worker
Two bugs surfaced after the fresh-wipe restart:

1. ROUND(double precision, integer) doesn't exist in Postgres. SQLite
   accepted ROUND(x::real, 1) but Postgres requires ::numeric. Cast
   every 2-arg ROUND in the API routes (uptime, analytics, routing-
   stats, trend). The broken queries made /api/uptime, /api/analytics,
   /api/routing-stats, /api/trend return 500 and also fed garbage into
   the dashboard.

2. The shared undici upstreamAgent's 30s keep-alive window caused
   bulk worker health pings to reuse dead sockets that providers had
   silently closed, surfacing as spurious TypeError: fetch failed and
   knocking ~100 good models into cooldown on every scan cycle. Drop
   dispatcher: upstreamAgent from worker/health.ts and remove the
   dangling imports in worker/scanner.ts and worker/benchmark.ts.
   Node's default fetch creates fresh connections, which is fine for
   the sparse worker path (~100 pings/hour). The chat completions hot
   path keeps the shared agent where keep-alive actually helps.

After deploy the operator should clear any lingering fetch-failed
cooldown rows:

  DELETE FROM health_logs WHERE error LIKE '%fetch failed%';

+++ src/app/api/analytics/route.ts
@@ -13,10 +13,10 @@ export async function GET() {
         COUNT(*) as total,
         SUM(CASE WHEN status >= 200 AND status < 300 THEN 1 ELSE 0 END) as success,
         ROUND(
-          SUM(CASE WHEN status >= 200 AND status < 300 THEN 1.0 ELSE 0.0 END) / COUNT(*) * 100,
+          (SUM(CASE WHEN status >= 200 AND status < 300 THEN 1.0 ELSE 0.0 END)::numeric / COUNT(*) * 100),
           1
         ) as "successRate",
-        ROUND(AVG(latency_ms)) as "avgLatencyMs"
+        ROUND(AVG(latency_ms)::numeric) as "avgLatencyMs"
       FROM gateway_logs
       WHERE created_at >= now() - interval '24 hours'
       GROUP BY provider
@@ -46,7 +46,7 @@ export async function GET() {
         COALESCE(resolved_model, request_model) as model,
         provider,
         COUNT(*) as count,
-        ROUND(AVG(latency_ms)) as "avgLatencyMs"
+        ROUND(AVG(latency_ms)::numeric) as "avgLatencyMs"
       FROM gateway_logs
       WHERE created_at >= now() - interval '24 hours'
       GROUP BY COALESCE(resolved_model, request_model), provider
